requirements.txt
torch
torchvision
pycocotools
grad-cam
requests
tqdm
scikit-learn
pillow
matplotlib
seaborn
tensorboard
